requirements.txt
anyio==4.10.0
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.3
Flask==3.1.2
Flask-SocketIO==5.5.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymodbus==3.11.1
PyMySQL==1.1.2
pyserial==3.5
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing_extensions==4.15.0
uvicorn==0.35.0
Werkzeug==3.1.3
wsproto==1.2.0
